requirements.txt
# Core Django Framework
Django==4.2.16
django-environ==0.11.2
django-cors-headers==4.3.1
django-ratelimit==4.1.0

# Database
psycopg2-binary==2.9.9
django-model-utils==4.3.1

# Authentication & Security
django-oauth-toolkit==1.7.1
PyJWT==2.8.0
cryptography==41.0.8
django-otp==1.5.0
qrcode==7.4.2

# Background Tasks
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# API & Validation
djangorestframework==3.14.0
pydantic==2.5.2
django-filter==23.5

# AWS Integration
boto3==1.34.32
botocore==1.34.32
aws-finops-dashboard==1.0.0

# Web Server & Production
gunicorn==21.2.0
whitenoise==6.6.0

# Utilities
requests==2.31.0
python-decouple==3.8
Pillow==10.1.0

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
coverage==7.3.2

# Monitoring & Logging
structlog==23.2.0
django-prometheus==2.3.1

# Security Testing
bandit==1.7.5
safety==2.3.5
